Move all serde impls into own module
