+++ p2panda-core/src/serde.rs
@@ -1,7 +1,15 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+use std::fmt;
+
+use serde::de::{Error as SerdeError, SeqAccess, Visitor};
+use serde::ser::SerializeSeq;
 use serde::{Deserialize, Serialize};
 use serde_bytes::{ByteBuf as SerdeByteBuf, Bytes as SerdeBytes};
+
+use crate::hash::{Hash, HashError};
+use crate::identity::{IdentityError, PrivateKey, PublicKey, Signature};
+use crate::operation::{Body, Header};
 
 /// Helper method for `serde` to serialize bytes into a hex string when using a human readable
 /// encoding (JSON, GraphQL), otherwise it serializes the bytes directly (CBOR).
@@ -30,9 +38,250 @@ where
     }
 }
 
+impl Serialize for Hash {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serialize_hex(self.as_bytes(), serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for Hash {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let bytes = deserialize_hex(deserializer)?;
+
+        bytes
+            .as_slice()
+            .try_into()
+            .map_err(|err: HashError| serde::de::Error::custom(err.to_string()))
+    }
+}
+
+impl Serialize for PrivateKey {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serialize_hex(self.as_bytes(), serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for PrivateKey {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let bytes = deserialize_hex(deserializer)?;
+
+        bytes
+            .as_slice()
+            .try_into()
+            .map_err(|err: IdentityError| serde::de::Error::custom(err.to_string()))
+    }
+}
+
+impl Serialize for PublicKey {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serialize_hex(self.as_bytes(), serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for PublicKey {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let bytes = deserialize_hex(deserializer)?;
+
+        bytes
+            .as_slice()
+            .try_into()
+            .map_err(|err: IdentityError| serde::de::Error::custom(err.to_string()))
+    }
+}
+
+impl Serialize for Signature {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serialize_hex(&self.to_bytes(), serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for Signature {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let bytes = deserialize_hex(deserializer)?;
+
+        bytes
+            .as_slice()
+            .try_into()
+            .map_err(|err: IdentityError| serde::de::Error::custom(err.to_string()))
+    }
+}
+
+impl Serialize for Header {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let mut seq = serializer.serialize_seq(None)?;
+        seq.serialize_element(&self.version)?;
+        seq.serialize_element(&self.public_key)?;
+
+        if let Some(signature) = &self.signature {
+            seq.serialize_element(signature)?;
+        }
+
+        seq.serialize_element(&self.payload_size)?;
+        seq.serialize_element(&self.payload_hash)?;
+        seq.serialize_element(&self.timestamp)?;
+        seq.serialize_element(&self.seq_num)?;
+
+        if let Some(backlink) = &self.backlink {
+            seq.serialize_element(backlink)?;
+        }
+
+        seq.serialize_element(&self.previous)?;
+
+        seq.end()
+    }
+}
+
+impl<'de> Deserialize<'de> for Header {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct HeaderVisitor;
+
+        impl<'de> Visitor<'de> for HeaderVisitor {
+            type Value = Header;
+
+            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                formatter.write_str("Header encoded as a sequence")
+            }
+
+            fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+            where
+                A: SeqAccess<'de>,
+            {
+                let version: u64 = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid version, expected u64"))?
+                    .ok_or(SerdeError::custom("version missing"))?;
+
+                let public_key: PublicKey = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid public key, expected bytes"))?
+                    .ok_or(SerdeError::custom("public key missing"))?;
+
+                let signature: Signature = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid signature, expected bytes"))?
+                    .ok_or(SerdeError::custom("signature missing"))?;
+
+                let payload_size: u64 = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid payload size, expected u64"))?
+                    .ok_or(SerdeError::custom("payload size missing"))?;
+
+                let payload_hash: Option<Hash> = match payload_size {
+                    0 => None,
+                    _ => {
+                        let hash: Hash = seq
+                            .next_element()
+                            .map_err(|_| {
+                                SerdeError::custom("invalid payload hash, expected bytes")
+                            })?
+                            .ok_or(SerdeError::custom("payload hash missing"))?;
+                        Some(hash)
+                    }
+                };
+
+                let timestamp: u64 = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid timestamp, expected u64"))?
+                    .ok_or(SerdeError::custom("timestamp missing"))?;
+
+                let seq_num: u64 = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid sequence number, expected u64"))?
+                    .ok_or(SerdeError::custom("sequence number missing"))?;
+
+                let backlink: Option<Hash> = match seq_num {
+                    0 => None,
+                    _ => {
+                        let hash: Hash = seq
+                            .next_element()
+                            .map_err(|err| {
+                                SerdeError::custom(format!(
+                                    "invalid backlink, expected bytes {err}"
+                                ))
+                            })?
+                            .ok_or(SerdeError::custom("backlink missing"))?;
+                        Some(hash)
+                    }
+                };
+
+                let previous: Vec<Hash> = seq
+                    .next_element()
+                    .map_err(|_| SerdeError::custom("invalid previous links, expected array"))?
+                    .ok_or(SerdeError::custom("previous array missing"))?;
+
+                Ok(Header {
+                    version,
+                    public_key,
+                    signature: Some(signature),
+                    payload_hash,
+                    payload_size,
+                    timestamp,
+                    seq_num,
+                    backlink,
+                    previous,
+                })
+            }
+        }
+
+        deserializer.deserialize_seq(HeaderVisitor)
+    }
+}
+
+impl Serialize for Body {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        serialize_hex(&self.0, serializer)
+    }
+}
+
+impl<'de> Deserialize<'de> for Body {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let bytes = deserialize_hex(deserializer)?;
+        Ok(Body(bytes.to_vec()))
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use serde::{Deserialize, Serialize};
+
+    use crate::hash::Hash;
+    use crate::identity::PublicKey;
 
     use super::{deserialize_hex, serialize_hex};
 
@@ -60,5 +309,69 @@ mod tests {
         // readable encoding
         let test: Test = ciborium::de::from_reader(&bytes[..]).unwrap();
         assert_eq!(test.0, vec![1, 2, 3]);
+    }
+
+    #[test]
+    fn serialize_hash() {
+        // Serialize CBOR (non human-readable byte encoding)
+        let mut bytes: Vec<u8> = Vec::new();
+        let hash = Hash::new([1, 2, 3]);
+        ciborium::ser::into_writer(&hash, &mut bytes).unwrap();
+        assert_eq!(
+            bytes,
+            vec![
+                88, 32, 177, 119, 236, 27, 242, 109, 251, 59, 112, 16, 212, 115, 230, 212, 71, 19,
+                178, 155, 118, 91, 153, 198, 230, 14, 203, 250, 231, 66, 222, 73, 101, 67
+            ]
+        );
+
+        // Serialize JSON (human-readable hex encoding)
+        let json = serde_json::to_string(&hash).unwrap();
+        assert_eq!(
+            json,
+            "\"b177ec1bf26dfb3b7010d473e6d44713b29b765b99c6e60ecbfae742de496543\""
+        );
+    }
+
+    #[test]
+    fn deserialize_hash() {
+        // Deserialize CBOR (non human-readable byte encoding)
+        let bytes = [
+            88, 32, 177, 119, 236, 27, 242, 109, 251, 59, 112, 16, 212, 115, 230, 212, 71, 19, 178,
+            155, 118, 91, 153, 198, 230, 14, 203, 250, 231, 66, 222, 73, 101, 67,
+        ];
+        let hash: Hash = ciborium::de::from_reader(&bytes[..]).unwrap();
+        assert_eq!(hash, Hash::new([1, 2, 3]));
+
+        // Deserialize JSON (human-readable hex encoding)
+        let json = "\"b177ec1bf26dfb3b7010d473e6d44713b29b765b99c6e60ecbfae742de496543\"";
+        let hash: Hash = serde_json::from_str(json).unwrap();
+        assert_eq!(hash, Hash::new([1, 2, 3]));
+    }
+
+    #[test]
+    fn serialize_public_key() {
+        // Serialize CBOR (non human-readable byte encoding)
+        let mut bytes: Vec<u8> = Vec::new();
+        let public_key = PublicKey::from_bytes(&[
+            215, 90, 152, 1, 130, 177, 10, 183, 213, 75, 254, 211, 201, 100, 7, 58, 14, 225, 114,
+            243, 218, 166, 35, 37, 175, 2, 26, 104, 247, 7, 81, 26,
+        ])
+        .unwrap();
+        ciborium::ser::into_writer(&public_key, &mut bytes).unwrap();
+        assert_eq!(
+            bytes,
+            vec![
+                88, 32, 215, 90, 152, 1, 130, 177, 10, 183, 213, 75, 254, 211, 201, 100, 7, 58, 14,
+                225, 114, 243, 218, 166, 35, 37, 175, 2, 26, 104, 247, 7, 81, 26,
+            ]
+        );
+
+        // Serialize JSON (human-readable hex encoding)
+        let json = serde_json::to_string(&public_key).unwrap();
+        assert_eq!(
+            json,
+            "\"d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511a\""
+        );
     }
 }
